Many changes. See commit details.
NSUserDefaults is no longer used to store syncing preferences. A hidden database file is used.
If not synchronizing with Dropbox, a local plist is used to store all the user preferences. It is placed in ~/Library/DBPreferences.
Settings are now automatically loaded from the appropriate path and are ready to be read as soon as the user creates a DBUserDefaults object.